requirements.txt
google-genai
python-dotenv
pandas
openpyxl
customtkinter
pyinstaller